+++ src/cpp/session/modules/SessionRCompletions.R
@@ -1609,20 +1609,6 @@ assign(x = ".rs.acCompletionTypes",
       {
          allNames <- dollarNamesMethod(object, "")
          
-         # detect completion systems that append closing parentheses
-         # to the completion item, and assume those are functions
-         # rJava and Rcpp will do this for some objects
-         # for example:
-         # - rJava:::.DollarNames.jobjref
-         # - Rcpp:::.DollarNames.C++Object
-         types <- ifelse(
-            grepl("[()]\\s*$", allNames),
-            .rs.acCompletionTypes$FUNCTION,
-            .rs.acCompletionTypes$UNKNOWN
-         )
-         allNames <- gsub("[()]*\\s*$", "", allNames)
-         attr(allNames, "types") <- as.integer(types)
-         
          # check for custom helpHandler
          helpHandler <- attr(allNames, "helpHandler", exact = TRUE)
       }
@@ -1715,15 +1701,24 @@ assign(x = ".rs.acCompletionTypes",
       else
       {
          type <- numeric(length(names))
+         looksLikeFunction <- grepl("[()]\\s*$", names, perl = TRUE)
          for (i in seq_along(names))
          {
-            type[[i]] <- suppressWarnings(tryCatch(
-               if (is.environment(object) && bindingIsActive(names[[i]], object))
-                  .rs.acCompletionTypes$UNKNOWN
-               else
-                  .rs.getCompletionType(eval(call("$", quote(object), names[[i]]))),
-               error = function(e) .rs.acCompletionTypes$UNKNOWN
-            ))
+            if (looksLikeFunction[[i]])
+            {
+               names[[i]] <- gsub("[()]+\\s*$", "", names[[i]], perl = TRUE)
+               type[[i]] <- .rs.acCompletionTypes$FUNCTION
+            }
+            else
+            {
+               type[[i]] <- suppressWarnings(tryCatch(
+                  if (is.environment(object) && bindingIsActive(names[[i]], object))
+                     .rs.acCompletionTypes$UNKNOWN
+                  else
+                     .rs.getCompletionType(eval(call("$", quote(object), names[[i]]))),
+                  error = function(e) .rs.acCompletionTypes$UNKNOWN
+               ))
+            }
          }
       }
    }