+++ src/cpp/tests/automation/testthat/test-automation-completions.R
@@ -122,3 +122,32 @@ test_that("autocompletions within piped expressions work at start of document", 
    })
 
 })
+
+# https://github.com/rstudio/rstudio/issues/15115
+test_that(".DollarNames completions still produce types", {
+   
+   remote$consoleExecuteExpr({
+   
+      className <- basename(tempfile(pattern = "class-"))
+      registerS3method(".DollarNames", className, function(x, pattern) names(x))
+      
+      . <- structure(
+         list(apple = identity, banana = identity),
+         class = className
+      )
+      
+   })
+   
+   completions <- remote$completionsRequest(".$")
+   expect_equal(completions, c("apple", "banana"))
+   
+   remote$consoleExecuteExpr({
+      className <- basename(tempfile(pattern = "class-"))
+      registerS3method(".DollarNames", className, function(x, pattern) c("example1()", "example2()"))
+      . <- structure(list(), class = className)
+   })
+   
+   completions <- remote$completionsRequest(".$")
+   expect_equal(completions, c("example1", "example2"))
+   
+})